0.0.97

+++ src/modules/date.js
@@ -57,7 +57,7 @@ this.date = (function () {
     date.julian.to.georgian = function (jd) {
         var wjd, depoch, quadricent, dqc, cent, dcent, quad, dquad,
             yindex, dyindex, year, month, day, yearday, leapadj;
-        jd = parseInt(jd);
+        jd = parseFloat(jd);
 
         wjd = Math.floor(jd - 0.5) + 0.5;
         depoch = wjd - GREGORIAN_EPOCH;
@@ -85,7 +85,7 @@ this.date = (function () {
     date.julian.to.persian = function (jd) {
         var year, month, day, depoch, cycle, cyear, ycycle,
             aux1, aux2, yday;
-        jd = parseInt(jd);
+        jd = parseFloat(jd);
 
         jd = Math.floor(jd) + 0.5;
